requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
dlib>=19.22.0
RPi.GPIO>=0.7.0

